PRICING-PAGE-SELL: per-tab + bundle buy-buttons on /pricing, honest partial-Stripe state
